move Fawks' SWAT to customSwat
