+++ addons/gear/loadouts/customSwat/crates/paramedic.hpp
