+++ addons/gear/loadouts/customSwat/crates/police.hpp
