+++ addons/gear/loadouts/customSwat/demo_explosive.hpp
