+++ addons/gear/loadouts/customSwat/demo_lead.hpp
