+++ addons/gear/loadouts/customSwat/ews_comp.hpp
