+++ addons/gear/loadouts/customSwat/loadout_base_Swat.hpp
