+++ addons/gear/loadouts/customSwat/lynx_driver.hpp
