+++ addons/gear/loadouts/customSwat/mtr_gunner.hpp
