+++ addons/gear/loadouts/customSwat/squad_aar.hpp
