+++ addons/gear/loadouts/customSwat/squad_ar.hpp
