+++ addons/gear/loadouts/customSwat/squad_lead.hpp
