+++ addons/gear/loadouts/customSwat/squad_marksman.hpp
