+++ addons/gear/loadouts/customSwat/squad_rifleman.hpp
